refactor: extracted class PersistedItem from sync to shorten class
