+++ consyncful.gemspec
@@ -37,7 +37,7 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'rubocop', '0.79.0'
 
   spec.add_dependency 'contentful', ['>=2.11.1', '<3.0.0']
+  spec.add_dependency 'hooks', '>=0.4.1'
   spec.add_dependency 'mongoid', ['>=7.0.2', '<8.0.0']
   spec.add_dependency 'rainbow'
-  spec.add_dependency 'hooks', '>=0.4.1'
 end
