+++ lib/consyncful/sync.rb
@@ -2,6 +2,7 @@
 
 require 'rainbow'
 require 'consyncful/item_mapper'
+require 'consyncful/persisted_item'
 require 'consyncful/stats'
 require 'hooks'
 
@@ -35,6 +36,7 @@ module Consyncful
 
     def run
       run_hook :before_run
+
       stats = Consyncful::Stats.new
       load_all_models
 
@@ -44,10 +46,9 @@ module Consyncful
 
       drop_stale
 
-      self.next_url = sync.next_sync_url
-      self.last_run_at = Time.current
-      save
+      update_run(sync.next_sync_url)
       stats.print_stats
+
       run_hook :after_run, changed_ids
     end
 
@@ -57,6 +58,12 @@ module Consyncful
       return unless defined? Rails
 
       Rails.application.eager_load!
+    end
+
+    def update_run(next_url)
+      self.next_url = next_url
+      self.last_run_at = Time.current
+      save
     end
 
     def start_sync
@@ -81,61 +88,8 @@ module Consyncful
 
     def sync_item(item, stats)
       puts Rainbow("syncing: #{item.id}").yellow
-      if item.deletion?
-        delete_model(item.id, stats)
-      else
-        create_or_update_model(item, stats)
-      end
+      PersistedItem.new(item, id, stats).persist
       item.id
-    end
-
-    def delete_model(id, stats)
-      Base.find_by(id: id).destroy
-      stats.record_deleted
-    rescue Mongoid::Errors::DocumentNotFound
-      puts Rainbow("Deleted record not found: #{id}").yellow
-      nil
-    end
-
-    def create_or_update_model(item, stats)
-      return if item.type.nil?
-
-      instance = find_or_initialize_item(item)
-      update_stats(instance, stats)
-
-      reset_fields(instance)
-
-      item.mapped_fields(DEFAULT_LOCALE).each do |field, value|
-        instance[field] = value
-      end
-
-      instance[:sync_id] = id
-
-      instance.save
-    end
-
-    def find_or_initialize_item(item)
-      model_class(item.type).find_or_initialize_by(id: item.id)
-    end
-
-    def update_stats(instance, stats)
-      if instance.persisted?
-        stats.record_updated
-      else
-        stats.record_added
-      end
-    end
-
-    def model_class(type)
-      Base.model_map[type] || Base
-    end
-
-    def reset_fields(instance)
-      instance.attributes.each do |field_name, _value|
-        next if field_name.in? %w[_id _type]
-
-        instance[field_name] = nil
-      end
     end
   end
 end
