+++ spec/consyncful/sync_spec.rb
@@ -35,12 +35,12 @@ RSpec.describe Consyncful::Sync do
 
     describe 'before run' do
       let(:callback) do
-        Proc.new { puts 'test before callback!' }
+        proc { puts 'test before callback!' }
       end
 
       it 'executes hook before run' do
         Consyncful::Sync.before_run callback
-        expect{sync.run}.to output(/\Atest before callback!/).to_stdout
+        expect { sync.run }.to output(/\Atest before callback!/).to_stdout
       end
 
       after do
@@ -60,7 +60,7 @@ RSpec.describe Consyncful::Sync do
       end
 
       let(:callback) do
-        Proc.new {|ids| puts "test after callback with #{ids.join(', ')}" }
+        proc { |ids| puts "test after callback with #{ids.join(', ')}" }
       end
 
       it 'executes hook after run and provides updated ids' do
@@ -110,61 +110,18 @@ RSpec.describe Consyncful::Sync do
       let(:page) do
         double('page', items: [double('item')])
       end
-      let(:mapper) { instance_double('Consyncful::ItemMapper') }
+      let(:mapper) { instance_double('Consyncful::ItemMapper', id: 'itemId') }
+      let(:persisted_item) { instance_double('Consyncful::PersistedItem') }
 
       before do
         allow(client_sync).to receive(:each_page).and_yield(page)
         allow(Consyncful::ItemMapper).to receive(:new).and_return(mapper)
       end
 
-      context 'when the item is a deletion' do
-        let(:mapper) { instance_double('Consyncful::ItemMapper', deletion?: true, id: 'itemId') }
-        let(:instance) { instance_double('Consyncful::Base') }
-
-        it 'finds the model and destroys it' do
-          expect(Consyncful::Base).to receive(:find_by).with(id: 'itemId').and_return(instance)
-          expect(instance).to receive(:destroy)
-          sync.run
-        end
-      end
-
-      context 'when the item is a create or update' do
-        let(:mapped_fields) { { field_name: 'value', other_field: 3 } }
-        let(:mapper) { instance_double('Consyncful::ItemMapper', deletion?: false, id: 'itemId', type: 'itemType', mapped_fields: mapped_fields) }
-        let(:instance) { instance_double('Consyncful::Base', persisted?: false, save: true, :[]= => nil, attributes: []) }
-
-        class self::TestContentfulType < Consyncful::Base
-          contentful_model_name 'itemType'
-        end
-        let(:klass) { self.class::TestContentfulType }
-
-        it 'finds the model of the correct type' do
-          expect(klass).to receive(:find_or_initialize_by).with(id: 'itemId').and_return(instance)
-          sync.run
-        end
-
-        it 'assigns all mapped fields' do
-          allow(klass).to receive(:find_or_initialize_by).with(id: 'itemId').and_return(instance)
-          expect(instance).to receive(:[]=).with(:field_name, 'value')
-          expect(instance).to receive(:[]=).with(:other_field, 3)
-          sync.run
-        end
-
-        it 'sets the sync id on the item' do
-          allow(klass).to receive(:find_or_initialize_by).with(id: 'itemId').and_return(instance)
-          expect(instance).to receive(:[]=).with(:sync_id, sync.id)
-          sync.run
-        end
-
-        it 'nils all other fields' do
-          skip 'need this test'
-        end
-
-        it 'saves the iteam' do
-          allow(klass).to receive(:find_or_initialize_by).with(id: 'itemId').and_return(instance)
-          expect(instance).to receive(:save)
-          sync.run
-        end
+      it 'persists the change' do
+        expect(Consyncful::PersistedItem).to receive(:new).with(mapper, sync.id, instance_of(Consyncful::Stats)).and_return(persisted_item)
+        expect(persisted_item).to receive(:persist)
+        sync.run
       end
     end
   end
